+++ nix/configuration.nix
@@ -14,12 +14,10 @@
     extra-substituters = [
       "https://pi.cachix.org"
       "https://nix-community.cachix.org"
-      "https://hyprland.cachix.org"
     ];
     extra-trusted-public-keys = [
       "pi.cachix.org-1:lGeoGJaZ5ZDabuRzkcD5EBTNnDM4HJ1vqeOxlWk1Flk="
       "nix-community.cachix.org-1:mB9FSh9qf2dCimDSUo8Zy7bkq5CX+/rkCWyvRCYg3Fs="
-      "hyprland.cachix.org-1:a7pgxzMz7+chwVL3/pzj6jIBMioiJM7ypFP8PwtkuGc="
     ];
   };
 
@@ -38,8 +36,6 @@
   virtualisation.docker.enable = true;
 
   programs.fish.enable = true;
-
-  programs.hyprland.enable = true;
 
   users.users.partkyle = {
     isNormalUser = true;
@@ -102,7 +98,6 @@
   xdg.portal = {
     enable = true;
     extraPortals = [
-      pkgs.xdg-desktop-portal-hyprland
       pkgs.xdg-desktop-portal-gtk
     ];
   };